backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
opencv-python==4.10.0.84
ultralytics==8.3.140
twilio==9.2.0
pyserial==3.5
python-dotenv==1.0.1
numpy==1.26.4
psycopg2-binary==2.9.9